Secured websocket interface

+++ nest/backend/src/gateways/message.gateway.ts
@@ -30,9 +30,15 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     private clients = new Map<number, Socket>();
     private sockets = new Map<Socket, number>();
 
+    private verifyId(client: Socket, id: number) {
+        if (this.sockets.get(client) != id)
+            throw new BadRequestException("Nice try");
+    }
+
     @SubscribeMessage('newmsg')
     async handleMessage(@MessageBody() dto: PostMessageDTO,
                         @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.senderId);
         const chan: Channel = await this.channelRepository.findOneBy({id: dto.channelId})
         if (!chan) throw new BadRequestException("Channel Doesn't exist")
         const user = chan.users.find(user => user.id === dto.senderId);
@@ -50,6 +56,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('join')
     async joinChannel(  @MessageBody() dto: JoinChannelDTO,
                         @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.uid);
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
         if (!chan) throw new BadRequestException("Channel Doesn't exist")
         const user = await this.userRepository.findOneBy({id: dto.uid});
@@ -64,6 +71,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('joinwithpw')
     async joinChannelWithPassword(  @MessageBody() dto: JoinChannelDTO,
                         @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.uid);
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
         if (!chan) throw new BadRequestException("Channel Doesn't exist")
         const user = await this.userRepository.findOneBy({id: dto.uid});
@@ -79,7 +87,8 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     async banUser(  @MessageBody() dto: BanUserDTO,
                     @ConnectedSocket() client: Socket) {
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
-        if (!chan) throw new BadRequestException("Channel Doesn't exist")
+        if (!chan) throw new BadRequestException("Channel Doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new BadRequestException("Nice try");
         const user = chan.users.find(user => user.id === dto.uid);
         if (!user) throw new BadRequestException("User doesn't exist or is not on this channel");
         if (dto.isBan) {
@@ -101,6 +110,8 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('getmsg')
     async getMessagePage(   @MessageBody() dto: GetMessageDTO,
                             @ConnectedSocket() client: Socket) {
+        if (!client.in(dto.channelId.toString()))
+            throw new BadRequestException("Nice try");
         const messages = await this.messageRepository.find({
             select: {
                 id:         true,
@@ -121,6 +132,9 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('unban')
     async unBanUser(@MessageBody() dto: JoinChannelDTO,
                     @ConnectedSocket() client: Socket) {
+        const chan = await this.channelRepository.findOneBy({id: dto.chanId})
+        if (!chan) throw new BadRequestException("Channel Doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new BadRequestException("Nice try");
         const ban = await this.bansAndMutesRepository.findOneBy({
             channel: {id: dto.chanId},
             user: {id: dto.uid}
@@ -134,6 +148,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
                         @ConnectedSocket() client: Socket) {
         const chan = await this.channelRepository.findOneBy({id})
         if (!chan) throw new BadRequestException("Channel doesn't exist");
+        if (chan.admin.id != this.sockets.get(client)) throw new BadRequestException("Nice try");
         this.server.to(chan.id.toString()).emit("/" + chan.id + ":Channel was deleted");
         this.server.socketsLeave(chan.id.toString());
         this.channelRepository.delete(id);
@@ -142,6 +157,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('create')
     async createChannel(@MessageBody() dto: CreateChannelDTO,
                         @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.adminId);
         const channel = this.channelRepository.create(dto);
         client.join(channel.id.toString());
         this.channelRepository.save(channel);
@@ -150,6 +166,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('leave')
     async leaveChannel( @MessageBody() dto: JoinChannelDTO,
                         @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.uid);
         const chan = await this.channelRepository.findOneBy({id: dto.chanId});
         client.leave(chan.id.toString());
         if (!chan || !chan.removeUser(dto.uid))
@@ -161,8 +178,10 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     async deleteMessage(@MessageBody() id: number,
                         @ConnectedSocket() client: Socket) {
         const message = await this.messageRepository.findOneBy({id});
+        const clientid = this.sockets.get(client);
         if (!message) throw new BadRequestException("Message doesn't exist");
-        if (message.sender.id )
+        if (message.sender.id != clientid && message.channel.admin.id != clientid)
+            throw new BadRequestException("Nice try");
         this.server.to(message.channel.id.toString()).emit("!" + id);
         this.messageRepository.delete(id)
         return HttpStatus.NO_CONTENT;
@@ -171,6 +190,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('sendDM')
     async sendDM(   @MessageBody() dto : SendDMDTO,
                     @ConnectedSocket() client: Socket) {
+        this.verifyId(client, dto.id1);
         const user1 = await this.userRepository.findOneBy({id: dto.id1});
         const user2 = await this.userRepository.findOneBy({id: dto.id2});
 
@@ -203,6 +223,7 @@ export class MessageGateway implements OnGatewayConnection, OnGatewayDisconnect,
     @SubscribeMessage('addFriend')
     async addFriend(@MessageBody() ids: number[],
                     @ConnectedSocket() client: Socket) {
+        this.verifyId(client, ids[0]);
         const user1 = await this.userRepository.findOneBy({id: ids[0]});
         const user2 = await this.userRepository.findOneBy({id: ids[1]});
